% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/football_api_url.R
\name{football_api_url}
\alias{football_api_url}
\title{Map name to base URL for API}
\usage{
football_api_url(api = "football-data.org")
}
\arguments{
\item{api}{Name of API. Currently has \code{"football-data.org"}}
}
\value{
Base url for API
}
\description{
Adding in as potentially may allow package to work over a few APIs
}
\examples{
football_api_url("football-data.org")
}
