% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/football_calculate_metrics.R
\name{football_calculate_competition_metrics}
\alias{football_calculate_competition_metrics}
\title{Take API data and calculate some metrics for fun}
\usage{
football_calculate_competition_metrics(competition_standings)
}
\arguments{
\item{competition_standings}{Data frame from \code{football_get_standing}}
}
\value{
For a particular competition - metrics on all teams.
}
\description{
Turn data into some fun 'how would a supporter be feeling' metrics.
}
