% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/football_get_standing.R
\name{football_get_standing}
\alias{football_get_standing}
\title{Map name to base URL for API}
\usage{
football_get_standing(
  competition_id,
  api = "football-data.org",
  api_token = Sys.getenv("API_FOOTBALL_DATA")
)
}
\arguments{
\item{competition_id}{The ID for the competition of interest.}

\item{api}{Name of API. Currently has \code{"football-data.org"}}

\item{api_token}{Your API token}
}
\value{
League standings as dataframe.
}
\description{
Adding in as potentially may allow package to work over a few APIs
}
