% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/football_get_upcoming.R
\name{football_get_upcoming}
\alias{football_get_upcoming}
\title{Get upcoming matches}
\usage{
football_get_upcoming(
  team_id = 86,
  api = "football-data.org",
  api_token = Sys.getenv("API_FOOTBALL_DATA")
)
}
\arguments{
\item{team_id}{What is the team id}

\item{api}{Which API to use.}

\item{api_token}{Your API token.}
}
\value{
Base url for API
}
\description{
Hits the matches API fixed on \code{status=SCHEDULED}.
}
\examples{
football_api_url()

}
