% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/football_pull.R
\name{football_pull}
\alias{football_pull}
\title{Internal function that does the pull}
\usage{
football_pull(
  api = "football-data.org",
  api_endpoint = "/competitions/",
  api_token = Sys.getenv("API_FOOTBALL_DATA")
)
}
\arguments{
\item{api}{Name of API. Currently has \code{"football-data.org"}}

\item{api_token}{}
}
\value{
Base url for API
}
\description{
Adding in as potentially may allow package to work over a few APIs,
and this step is done over multiple functions.
}
\examples{
# Will error as no API token
football_api_url("football-data.org")
}
